chore: 同步gerrit代码到github
chore: 修改插件编程指南

补充缺失的接口函数和消息协议，优化文档格式和内容

Log:
Influence: 修改文档，不涉及功能
(cherry picked from commit cb89e7ca3e63f72cbcd8b3958a256bd813ef9d53)

feat: 删除写入到cache里面的日志

锁屏的日志已经写入到journal了，无需向cache目录重复写入日志

Log: 日志不再写入到cache目录
Influence: 日志
Task: https://pms.uniontech.com/task-view-379327.html
(cherry picked from commit 4f21e8889ffd8939f803240c898bde278b047507)

feat: 控制台日志只在调试模式下输出

发布模式下不输出控制台日志

Log: 控制台日志只在调试模式下输出
Influence: 日志
Task: https://pms.uniontech.com/task-view-379327.html
(cherry picked from commit de95b99e757a2ef3331b46809dc55af2fef4f1b0)

fix: 【DDE】【终端锁定】开启自动登录和无密码登录，终端锁定后，登录按钮未置灰显示

  无密码认证时，需要同时考虑终端锁定的情况

Bug: https://pms.uniontech.com/bug-view-268089.html
Log: 修复开启自动登录和无密码登录，终端锁定后，登录按钮未置灰显示问题
(cherry picked from commit 954e344d7e32ccd0ee6732635711cff01fd4257c)
(cherry picked from commit c59d73173a38b1cd7607c2bffde9565ad4a2822f)

fix: 待机信号处理中考虑其它流程设置的可见性

有些流程需要内置黑屏，所以可能在未锁定时拉起lock

Log: 响应唤醒时，处理可见性
Bug: https://pms.uniontech.com/bug-view-314491.html
(cherry picked from commit eb18a13d0a2b1a18e30fa5785dda72093bad4b8f)
(cherry picked from commit e32cd5b00206307a187be1b0c77e96aa380b2173)

feat: 切换用户界面待机唤醒时进入锁屏界面

切换用户界面属于进入锁屏，唤起时仍进入锁屏界面，需输入密码进入系统

Log:
Task: https://pms.uniontech.com/task-view-379555.html
Influence: 切换用户界面待机唤醒，以及各种和切换用户界面相关的逻辑
(cherry picked from commit eb1221425988cdb7469ff0fc479ba3927df0dd1d)

fix: 极端情况无法获取到后端账户User列表信息，将页面显示成输入密码和账户的情况

极端情况无法获取到后端账户User列表信息，将页面显示成输入密码和账户的情况

Log: 极端情况无法获取到后端账户User列表信息，将页面显示成输入密码和账户的情况
Influence: 获取不到账户信息是greeter崩溃
Bug: https://pms.uniontech.com/bug-view-322207.html
(cherry picked from commit ebb3a2e3839821a759598a17c02fbb460595e865)

fix: 登录界面显示两个重置密码弹窗

auth_password新建后，旧的析构未有检测重置密码弹窗，如果有，则应该关闭和释放内存。

Log: as title
Influence: 重置密码弹窗显示一个
Bug: https://pms.uniontech.com/bug-view-255785.html
(cherry picked from commit e44decfc7c1aa030a721059880955c41f34a0810)

fix: 如果开启了Checkpwd配置，关机时不能显示关机黑屏

如果开启了Checkpwd配置，关机时不能显示关机黑屏
以前增加关机黑屏，是为了解决关机、重启闪屏，但是未考虑到关机还有一个gsettings配置checkpwd

Log: 如果开启了Checkpwd配置，关机时不能显示关机黑屏
Influence: 开启配置关机
Bug: https://pms.uniontech.com/bug-view-328319.html
(cherry picked from commit 5df869bafb8c0a1022161610e8f343a1ccc50bc9)

fix: 修复密码错误详细弹框UI问题

1. 密码错误详细弹框UI高度过高

Log: 高倍缩放（大于1倍）下，登录或锁屏界面下，密码错误的tooltip提示显示异常
Influence: 密码错误详细弹框
Bug: https://pms.uniontech.com/bug-view-328583.html
(cherry picked from commit c8d180c73cb525affe499114939a5985a1ef66e0)

chore: 避免频繁通过接口查询错误提示框使能状态

没有必要的接口频繁调用

(cherry picked from commit 6d30c6b4a51886c7ab4fe822c1f8eabe56a27663)

fix: 重新修改免密登录待机唤醒场景

1. 锁屏界面待机时，唤醒进入锁屏免密界面
2. 关机界面待机时，唤醒进入桌面

Log: 【通用】关闭待机恢复时需要密码开关，开启免密登录开关，在桌面待机唤醒后，显示锁屏界面，没有直接进入桌面
Influence: 免密待机唤醒
Bug: https://pms.uniontech.com/bug-view-330173.html
(cherry picked from commit 59c4c850fd0faab07f1482e6cb784c4c4ef5ddb3)

fix: 处理缩放和当前屏幕不匹配的情况

启动greeter之前判断下从配置文件中获取缩放和当前屏幕是否匹配，如果不匹配则用默认缩放1
这个只能处理启动greeter前插入屏幕的场景，greeter启动后插入和当前缩放不匹配的屏幕无法处理。

Log:
Bug: https://pms.uniontech.com/bug-view-331209.html
Influence: 登录界面缩放（设计多种缩放比，最好用4K屏幕测试），多屏缩放比（相同分辨率屏幕和不同分辨率屏幕都需要测试）
(cherry picked from commit 813f345c203ad4b47356c10367a923e379bf65c8)

chore: 手势插件安装问题

明确dconfig配置文件归属，避免产生安装时文件冲突问题

(cherry picked from commit d09727b0745c92f18cb3d445e6bb9069ebb232b8)

fix: 修复pgy待机唤醒过程中mouseRelease造成锁屏隐藏的问题

触发了WarningContent的展示，后响应鼠标事件隐藏锁屏

Log: 调整显示流程
Influence: 修复问题
Bug: https://pms.uniontech.com/bug-view-331651.html
(cherry picked from commit 31cf025ab5d3cd29cd9b4027a5d42068c148b352)

fix: 修复唤醒过程中频繁点击造成锁屏隐藏的问题

存在两个方面的影响：
1 UI设置不正确（显示WarningContent)
2 外部设置的lock状态导致判断条件不满足

Log: 调整逻辑
Influence: 修复问题
Bug: https://pms.uniontech.com/bug-view-331651.html
(cherry picked from commit a0368c0fd23021c7b0bfba47e9f1b185f08c7a87)

chore: 增加显示登录框的日志

增加显示登录框的日志，用以后面分析问题

Log:
Influence: 无功能影响
(cherry picked from commit 0a62937301974e8ad92222b9f0559d0253b2a51d)

+++ src/dde-lock/lockworker.cpp
@@ -190,27 +190,39 @@ void LockWorker::initConnections()
         const bool sleepLock = isSleepLock();
 #endif
         qCInfo(DDE_SHELL) << "Lock screen when system wakes up: " << sleepLock << ", is visible:" << m_model->visible();
+
+        FullScreenBackground::setContent(LockContent::instance());
+        m_model->setCurrentContentType(SessionBaseModel::LockContent);
+
         if (isSleep) {
-            endAuthentication(m_account, AT_All);
-            destroyAuthentication(m_account);
             checkSystemWakeupTimer->start();
             doSuspendTime = QDateTime::currentDateTime();
+            m_model->setIsBlackMode(true);
+            // 休眠时按需拉起锁屏,注意此时是被动响应休眠
+            if (sleepLock && m_login1SessionSelf->active()) {
+                // 仅sleepLock配置开启时执行
+                m_model->setCurrentModeState(SessionBaseModel::ModeStatus::PasswordMode);
+                setLocked(true);
+                m_model->setVisible(true);
+
+                qCInfo(DDE_SHELL) << "locked for sleepLock config";
+            }
         } else {
-            // 非黑屏mode
-            m_model->setIsBlackMode(isSleep);
-
-            // 如果待机唤醒后需要密码则创建验证
-            if(m_login1SessionSelf->active() && sleepLock)
-                createAuthentication(m_model->currentUser()->name());
-        }
-        if (!m_model->visible() && sleepLock) {
-            m_model->setIsBlackMode(isSleep);
-            m_model->setVisible(true);
-        }
-
-        if (!isSleep && !sleepLock) {
-            //待机唤醒后检查是否需要密码，若不需要密码直接隐藏锁定界面
-            m_model->setVisible(false);
+            SessionBaseModel::ModeStatus status = m_model->currentModeState();
+            if (isLocked() || (m_model->isNoPasswordLogin() && (status == SessionBaseModel::ModeStatus::PowerMode || status == SessionBaseModel::ModeStatus::PasswordMode))) {
+                // 唤醒时如果处于lock状态，重置认证
+                if (m_login1SessionSelf->active()) {
+                    endAuthentication(m_account, AT_All);
+                    destroyAuthentication(m_account);
+                    createAuthentication(m_model->currentUser()->name());
+                }
+            } else {
+                // 处理其他流程设置的可见状态
+                m_model->setVisible(false);
+            }
+            // 避免在电源选项处理中设置密码模式，会导致唤醒后闪锁屏
+            m_model->setCurrentModeState(SessionBaseModel::ModeStatus::PasswordMode);
+            m_model->setIsBlackMode(false);
         }
         emit m_model->prepareForSleep(isSleep);
     });
@@ -286,40 +298,50 @@ void LockWorker::initConnections()
 
 void LockWorker::initData()
 {
-    /* com.deepin.daemon.Accounts */
-    m_model->updateUserList(m_accountsInter->userList());
-    m_model->updateLoginedUserList(m_loginedInter->userList());
+    auto list = m_accountsInter->userList();
+    if (!list.isEmpty()) {
+        /* com.deepin.daemon.Accounts */
+        m_model->updateUserList(m_accountsInter->userList());
+        m_model->updateLoginedUserList(m_loginedInter->userList());
 
-    m_model->setUserlistVisible(valueByQSettings<bool>("", "userlist", true));
-    /* com.deepin.udcp.iam */
-    QDBusInterface ifc(DSS_DBUS::udcpIamService, DSS_DBUS::udcpIamPath, DSS_DBUS::udcpIamService, QDBusConnection::systemBus(), this);
-    const bool allowShowCustomUser = (!m_model->userlistVisible()) || valueByQSettings<bool>("", "loginPromptInput", false) ||
-        ifc.property("Enable").toBool() || checkIsADDomain();
-    m_model->setAllowShowCustomUser(allowShowCustomUser);
+        m_model->setUserlistVisible(valueByQSettings<bool>("", "userlist", true));
+        /* com.deepin.udcp.iam */
+        QDBusInterface ifc(DSS_DBUS::udcpIamService, DSS_DBUS::udcpIamPath, DSS_DBUS::udcpIamService, QDBusConnection::systemBus(), this);
+        const bool allowShowCustomUser = (!m_model->userlistVisible()) || valueByQSettings<bool>("", "loginPromptInput", false) ||
+            ifc.property("Enable").toBool() || checkIsADDomain();
+        m_model->setAllowShowCustomUser(allowShowCustomUser);
 
-    /* init server user or custom user */
-    if (DSysInfo::deepinType() == DSysInfo::DeepinServer || m_model->allowShowCustomUser()) {
-        std::shared_ptr<User> user(new User());
-        m_model->setIsServerModel(DSysInfo::deepinType() == DSysInfo::DeepinServer);
-        m_model->addUser(user);
-    }
+        /* init server user or custom user */
+        if (DSysInfo::deepinType() == DSysInfo::DeepinServer || m_model->allowShowCustomUser()) {
+            std::shared_ptr<User> user(new User());
+            m_model->setIsServerModel(DSysInfo::deepinType() == DSysInfo::DeepinServer);
+            m_model->addUser(user);
+        }
 
-    /* com.deepin.dde.LockService */
-    std::shared_ptr<User> user_ptr = m_model->findUserByUid(getuid());
-    if (user_ptr.get()) {
-        m_model->updateCurrentUser(user_ptr);
-        const QString &userJson = m_lockInter->CurrentUser();
-        QJsonParseError jsonParseError;
-        const QJsonDocument userDoc = QJsonDocument::fromJson(userJson.toUtf8(), &jsonParseError);
-        if (jsonParseError.error != QJsonParseError::NoError || userDoc.isEmpty()) {
-            qCWarning(DDE_SHELL) << "Failed to obtain current user information from lock service!";
+        /* com.deepin.dde.LockService */
+        std::shared_ptr<User> user_ptr = m_model->findUserByUid(getuid());
+        if (user_ptr.get()) {
+            m_model->updateCurrentUser(user_ptr);
+            const QString &userJson = m_lockInter->CurrentUser();
+            QJsonParseError jsonParseError;
+            const QJsonDocument userDoc = QJsonDocument::fromJson(userJson.toUtf8(), &jsonParseError);
+            if (jsonParseError.error != QJsonParseError::NoError || userDoc.isEmpty()) {
+                qCWarning(DDE_SHELL) << "Failed to obtain current user information from lock service!";
+            } else {
+                const QJsonObject userObj = userDoc.object();
+                m_model->currentUser()->setLastAuthType(AUTH_TYPE_CAST(userObj["AuthType"].toInt()));
+                m_model->currentUser()->setLastCustomAuth(userObj["LastCustomAuth"].toString());
+            }
         } else {
-            const QJsonObject userObj = userDoc.object();
-            m_model->currentUser()->setLastAuthType(AUTH_TYPE_CAST(userObj["AuthType"].toInt()));
-            m_model->currentUser()->setLastCustomAuth(userObj["LastCustomAuth"].toString());
+            m_model->updateCurrentUser(m_lockInter->CurrentUser());
         }
     } else {
-        m_model->updateCurrentUser(m_lockInter->CurrentUser());
+        qCWarning(DDE_SHELL) << "dbus com.deepin.daemon.Accounts userList is empty, use ...";
+
+        std::shared_ptr<User> user(new User());
+        m_model->addUser(user);
+        m_model->updateCurrentUser(user);
+        m_model->setAllowShowCustomUser(true);
     }
 
     /* com.deepin.daemon.Authenticate */
@@ -342,6 +364,8 @@ void LockWorker::initConfiguration()
     m_model->setAlwaysShowUserSwitchButton(getDconfigValue("switchUser", Ondemand).toInt() == AuthInterface::Always);
     m_model->setAllowShowUserSwitchButton(getDconfigValue("switchUser", Ondemand).toInt() == AuthInterface::Ondemand);
 #endif
+
+    m_model->setGsCheckpwd(isCheckPwdBeforeRebootOrShut());
 
     checkPowerInfo();
 }
@@ -490,7 +514,6 @@ void LockWorker::doPowerAction(const SessionBaseModel::PowerAction action)
     case SessionBaseModel::PowerAction::RequireSuspend:
     {
         m_model->setIsBlackMode(true);
-        m_model->setCurrentModeState(SessionBaseModel::ModeStatus::PasswordMode);
 
         int delayTime = 500;
 #ifndef ENABLE_DSS_SNIPE
@@ -520,7 +543,6 @@ void LockWorker::doPowerAction(const SessionBaseModel::PowerAction action)
     case SessionBaseModel::PowerAction::RequireHibernate:
     {
         m_model->setIsBlackMode(true);
-        m_model->setCurrentModeState(SessionBaseModel::ModeStatus::PasswordMode);
 
         int delayTime = 500;
 #ifndef ENABLE_DSS_SNIPE
@@ -549,42 +571,44 @@ void LockWorker::doPowerAction(const SessionBaseModel::PowerAction action)
         }
     }
         break;
-    case SessionBaseModel::PowerAction::RequireRestart:
+    case SessionBaseModel::PowerAction::RequireRestart: {
         m_model->setShutdownMode(true);
+        auto gsCheckPwd = m_model->gsCheckpwd();
+        if (!isLocked() || m_model->currentModeState() == SessionBaseModel::ModeStatus::ShutDownMode || !gsCheckPwd) {
+            m_sessionManagerInter->RequestReboot();
+        } else {
+            createAuthentication(m_account);
+            m_model->setCurrentModeState(SessionBaseModel::ModeStatus::ConfirmPasswordMode);
+        }
 
-        QTimer::singleShot(350, this, [=] {
-            if (!isLocked() || m_model->currentModeState() == SessionBaseModel::ModeStatus::ShutDownMode || !isCheckPwdBeforeRebootOrShut()) {
-                m_sessionManagerInter->RequestReboot();
-            } else {
-                createAuthentication(m_account);
-                m_model->setCurrentModeState(SessionBaseModel::ModeStatus::ConfirmPasswordMode);
-            }
+        if (m_model->visibleShutdownWhenRebootOrShutdown()) {
+            return;
+        }
 
-            if (m_model->visibleShutdownWhenRebootOrShutdown()) {
-                return;
-            }
-
+        if (!gsCheckPwd)
             m_model->setVisible(false);
-        });
+
         return;
-    case SessionBaseModel::PowerAction::RequireShutdown:
+    }
+    case SessionBaseModel::PowerAction::RequireShutdown: {
         m_model->setShutdownMode(true);
+        auto gsCheckPwd = m_model->gsCheckpwd();
+        if (!isLocked() || m_model->currentModeState() == SessionBaseModel::ModeStatus::ShutDownMode || !gsCheckPwd) {
+            m_sessionManagerInter->RequestShutdown();
+        } else {
+            createAuthentication(m_account);
+            m_model->setCurrentModeState(SessionBaseModel::ModeStatus::ConfirmPasswordMode);
+        }
 
-        QTimer::singleShot(350, this, [=] {
-            if (!isLocked() || m_model->currentModeState() == SessionBaseModel::ModeStatus::ShutDownMode || !isCheckPwdBeforeRebootOrShut()) {
-                m_sessionManagerInter->RequestShutdown();
-            } else {
-                createAuthentication(m_account);
-                m_model->setCurrentModeState(SessionBaseModel::ModeStatus::ConfirmPasswordMode);
-            }
+        if (m_model->visibleShutdownWhenRebootOrShutdown()) {
+            return;
+        }
 
-            if (m_model->visibleShutdownWhenRebootOrShutdown()) {
-                return;
-            }
-
+        if (!gsCheckPwd)
             m_model->setVisible(false);
-        });
+
         return;
+    }
     case SessionBaseModel::PowerAction::RequireLock:
         m_model->setCurrentModeState(SessionBaseModel::ModeStatus::PasswordMode);
         createAuthentication(m_model->currentUser()->name());
